+++ app/Dockerfile
@@ -11,4 +11,4 @@ RUN pip install --upgrade pip
 RUN pip install -r requirements.txt
 
 #CMD flask run --host=0.0.0.0 --port=5000
-CMD ["python", "app.py"]
+CMD ["python", "main.py"]
